AAindex: Amino Acid Index Database

A collection of published indices, or scales, of numerous physicochemical and biological properties of the 20 standard aminoacids (Release 9.1, August 2006).

data(aa.index)

Format

A list of 544 named indeces each with the following components:

H character vector: Accession number.

D character vector: Data description.

R character vector: LITDB entry number.

A character vector: Author(s).

T character vector: Title of the article.

J character vector: Journal reference.

C named numeric vector: Correlation coefficients of similar indeces (with coefficients of 0.8/-0.8 or more/less). The correlation coefficient is calculated with zeros filled for missing values.

I named numeric vector: Amino acid index data.

Source

‘AAIndex’ was obtained from:
http://www.genome.jp/aaindex/
For a description of the ‘AAindex’ database see:
http://www.genome.jp/aaindex/aaindex_help.html.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

‘AAIndex’ is the work of Kanehisa and co-workers:
Kawashima and Kanehisa (2000) Nucleic Acids Res. 28, 374;
Tomii and Kanehisa (1996) Protein Eng. 9, 27--36;
Nakai, Kidera and Kanehisa (1988) Protein Eng. 2, 93--100.

Examples

## Load AAindex data
data(aa.index)

## Find all indeces described as "volume"
ind <- which(sapply(aa.index, function(x)
                    length(grep("volume", x$D, ignore.case=TRUE)) != 0))

## find all indeces with author "Kyte"
ind <- which(sapply(aa.index, function(x) length(grep("Kyte", x$A)) != 0))

## examine the index
aa.index[[ind]]$I

## find indeces which correlate with it
all.ind <- names(which(Mod(aa.index[[ind]]$C) >= 0.88))

## examine them all
sapply(all.ind, function (x) aa.index[[x]]$I)