Table of Relevant Amino Acids

This data set provides the atomic masses of a selection of amino acids regularly occuring in proteins.

aa.table

Format

A data frame with the following components.

aa3

a character vector containing three-letter amino acid code.

aa1

a character vector containing one-letter amino acid code.

mass

a numeric vector containing the mass of the respective amino acids.

formula

a character vector containing the formula of the amino acid in which the mass calculat was based.

name

a character vector containing the full names of the respective amino acids.

See also

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Examples

data(aa.table)
aa.table

#>     aa3 aa1    mass       formula                             name
#> ALA ALA   A  71.078    C3 H5 N O1                          Alanine
#> ARG ARG   R 157.194  C6 H13 N4 O1                         Arginine
#> ASN ASN   N 114.103   C4 H6 N2 O2                       Asparagine
#> ASP ASP   D 114.079    C4 H4 N O3                    Aspartic Acid
#> CYS CYS   C 103.143  C3 H5 N O1 S                          Cystein
#> GLN GLN   Q 117.126   C4 H9 N2 O2                        Glutamine
#> GLU GLU   E 128.106    C5 H6 N O3                    Glutamic Acid
#> GLY GLY   G  57.051    C2 H3 N O1                          Glycine
#> HIS HIS   H 137.139   C6 H7 N3 O1                        Histidine
#> ILE ILE   I 113.158   C6 H11 N O1                       Isoleucine
#> LEU LEU   L 113.158   C6 H11 N O1                          Leucine
#> LYS LYS   K 129.180  C6 H13 N2 O1                           Lysine
#> MET MET   M 131.196  C5 H9 N O1 S                       Methionine
#> PHE PHE   F 147.174    C9 H9 N O1                    Phenylalanine
#> PRO PRO   P  97.115    C5 H7 N O1                          Proline
#> SER SER   S  87.077    C3 H5 N O2                           Serine
#> THR THR   T 101.104    C4 H7 N O2                        Threonine
#> TRP TRP   W 186.210 C11 H10 N2 O1                       Tryptophan
#> TYR TYR   Y 163.173    C9 H9 N O2                         Tyrosine
#> VAL VAL   V  99.131    C5 H9 N O1                           Valine
#> ABA ABA   X  85.104   C4 H7 N1 O1          alpha-aminobutyric acid
#> ASH ASH   D 115.087    C4 H5 N O3            Aspartic acid Neutral
#> CIR CIR   R 157.170  C6 H11 N3 O2                       citrulline
#> CME CME   C 179.260 C5 H9 N O2 S2 s,s-(2-hydroxyethyl)thiocysteine
#> CMT CMT   C 115.154  C4 H5 N O1 S                 o-methylcysteine
#> CSD CSD   C 134.134  C3 H4 N O3 S          s-cysteinesulfinic acid
#> CSO CSO   C 119.142  C3 H5 N O2 S                s-hydroxycysteine
#> CSW CSW   C 135.142  C3 H5 N O3 S               cysteine-s-dioxide
#> CSX CSX   C 119.142  C3 H5 N O2 S                   s-oxy cysteine
#> CYM CYM   C 102.135  C3 H4 N O1 S                 Cystein Negative
#> CYX CYX   C 102.135  C3 H4 N O1 S                   Cystein SSbond
#> DDE DDE   H 280.346 C13 H22 N5 O2                      diphthamide
#> GLH GLH   E 129.114    C5 H7 N O3           Glutatmic acid Neutral
#> HID HID   H 137.139   C6 H7 N3 O1                        Histidine
#> HIE HIE   H 137.139   C6 H7 N3 O1                        Histidine
#> HIP HIP   H 138.147   C6 H8 N3 O1               Histidine Positive
#> HSD HSD   H 137.139   C6 H7 N3 O1                        Histidine
#> HSE HSE   H 137.139   C6 H7 N3 O1                        Histidine
#> HSP HSP   H 138.147   C6 H8 N3 O1               Histidine Positive
#> IAS IAS   D 115.087    C4 H5 N O3                    beta-aspartyl
#> KCX KCX   K 172.182  C7 H12 N2 O3        lysine nz-carboxylic acid
#> LYN LYN   K 129.180  C6 H13 N2 O1                   Lysine Neutral
#> MHO MHO   M 147.195  C5 H9 N O2 S                  s-oxymethionine
#> MLY MLY   K 156.225  C8 H16 N2 O1                n-dimethyl-lysine
#> MSE MSE   M 178.091 C5 H9 N O1 SE                 selenomethionine
#> OCS OCS   C 151.141  C3 H5 N O4 S            cysteinesulfonic acid
#> PFF PFF   F 165.164  C9 H8 F N O1         4-fluoro-l-phenylalanine
#> PTR PTR   Y 243.153 C9 H10 N O5 P                o-phosphotyrosine
#> SEP SEP   S 167.057  C3 H6 N O5 P                    phosphoserine
#> TPO TPO   T 181.084  C4 H8 N O5 P                 phosphothreonine

## table look up
aa.table["HIS", ]

#>     aa3 aa1    mass     formula      name
#> HIS HIS   H 137.139 C6 H7 N3 O1 Histidine

## read PDB, and fetch residue masses
pdb <- read.pdb(system.file("examples/1hel.pdb", package="bio3d"))
aa2mass(pdb)

#>   [1] 130.1879  99.1310 147.1740  57.0510 157.1940 103.1430 128.1060 113.1580
#>   [9]  71.0780  71.0780  71.0780 131.1960 129.1800 157.1940 137.1390  57.0510
#>  [17] 113.1580 114.0790 114.1030 163.1730 157.1940  57.0510 163.1730  87.0770
#>  [25] 113.1580  57.0510 114.1030 186.2100  99.1310 103.1430  71.0780  71.0780
#>  [33] 129.1800 147.1740 128.1060  87.0770 114.1030 147.1740 114.1030 101.1040
#>  [41] 117.1260  71.0780 101.1040 114.1030 157.1940 114.1030 101.1040 114.0790
#>  [49]  57.0510  87.0770 101.1040 114.0790 163.1730  57.0510 113.1580 113.1580
#>  [57] 117.1260 113.1580 114.1030  87.0770 157.1940 186.2100 186.2100 103.1430
#>  [65] 114.1030 114.0790  57.0510 157.1940 101.1040  97.1150  57.0510  87.0770
#>  [73] 157.1940 114.1030 113.1580 103.1430 114.1030 113.1580  97.1150 103.1430
#>  [81]  87.0770  71.0780 113.1580 113.1580  87.0770  87.0770 114.0790 113.1580
#>  [89] 101.1040  71.0780  87.0770  99.1310 114.1030 103.1430  71.0780 129.1800
#>  [97] 129.1800 113.1580  99.1310  87.0770 114.0790  57.0510 114.1030  57.0510
#> [105] 131.1960 114.1030  71.0780 186.2100  99.1310  71.0780 186.2100 157.1940
#> [113] 114.1030 157.1940 103.1430 129.1800  57.0510 101.1040 114.0790  99.1310
#> [121] 117.1260  71.0780 186.2100 113.1580 157.1940  57.0510 103.1430 157.1940
#> [129] 130.1653