Bhattacharyya Coefficient

Calculate the Bhattacharyya Coefficient as a similarity between two modes objects.

bhattacharyya(...)

# S3 method for enma
bhattacharyya(enma, covs=NULL, ncore=NULL, ...)

# S3 method for array
bhattacharyya(covs, ncore=NULL, ...)

# S3 method for matrix
bhattacharyya(a, b, q=90, n=NULL, ...)

# S3 method for nma
bhattacharyya(...)

# S3 method for pca
bhattacharyya(...)

Arguments

enma
an object of class "enma" obtained from function nma.pdbs.

covs
an array of covariance matrices of equal dimensions.

ncore
number of CPU cores used to do the calculation. ncore>1 requires package ‘parallel’ installed.

a
covariance matrix to be compared with b.

b
covariance matrix to be compared with a.

q
a numeric value (in percent) determining the number of modes to be compared.

n
the number of modes to be compared.

...
arguments passed to associated functions.

Details

Bhattacharyya coefficient provides a means to compare two covariance matrices derived from NMA or an ensemble of conformers (e.g. simulation or X-ray conformers).

Value

Returns the similarity coefficient(s).

References

Skjaerven, L. et al. (2014) BMC Bioinformatics 15, 399. Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696. Fuglebakk, E. et al. (2013) JCTC 9, 5618--5628.

Author

Lars Skjaerven

See also