Obtain A SSE Object From An SSE Sequence Vector

Inverse process of the funciton pdb2sse.

bounds.sse(x, pdb = NULL)

Arguments

x
a character vector indicating SSE for each amino acid residue.

pdb
an object of class pdb as obtained from function read.pdb. Can be ignored if x has 'names' attribute for residue labels.

Value

a 'sse' object.

Details

call for its effects.

Note

In both $helix and $sheet, an additional $id component is added to indicate the original numbering of the sse. This is particularly useful in e.g. trim.pdb() function.

Author

Xin-Qiu Yao & Barry Grant

See also

Examples

# \donttest{
   # PDB server connection required - testing excluded

   pdb <- read.pdb("1a7l")

#>   Note: Accessing on-line PDB file

   sse <- pdb2sse(pdb)
   sse.ind <- bounds.sse(sse)
   sse.ind 

#> 
#> Call:
#>   NULL
#> 
#> Class:
#>   sse
#> 
#> Helices:  49 
#>         17-31 (A)      43-53 (A)      64-72 (A)      83-88 (A)      91-97 (A)
#>       132-142 (A)    154-163 (A)    186-200 (A)    210-218 (A)    229-238 (A)
#>       273-282 (A)    287-296 (A)    305-313 (A)    315-326 (A)    334-351 (A)
#>       357-365 (A)    134-137 (A)      17-31 (B)      43-50 (B)      64-73 (B)
#>         83-88 (B)      91-97 (B)    132-142 (B)    154-163 (B)    186-200 (B)
#>       210-219 (B)    229-238 (B)    273-282 (B)    287-296 (B)    305-313 (B)
#>       315-326 (B)    334-351 (B)    357-132 (B)      17-31 (C)      43-53 (C)
#>         64-72 (C)      83-88 (C)      91-97 (C)    132-142 (C)    154-163 (C)
#>       186-200 (C)    210-218 (C)    229-238 (C)    273-282 (C)    287-296 (C)
#>       305-313 (C)    315-326 (C)    334-352 (C)    357-363 (C)
#> 
#> Sheets:   24 
#>          7-10 (A)      35-38 (A)      59-63 (A)    260-266 (A)    106-111 (A)
#>       224-227 (A)    114-118 (A)    242-245 (A)       6-10 (B)      34-38 (B)
#>         59-63 (B)    260-266 (B)    106-111 (B)    224-227 (B)    114-118 (B)
#>       242-245 (B)       7-10 (C)      35-38 (C)      59-63 (C)    260-266 (C)
#>       106-111 (C)    224-227 (C)    114-118 (C)    242-245 (C)
#> 
#> Turns:   0 
#> 
#> Output is provided in residue numbers
#> 

# }