DCCM: Dynamical Cross-Correlation Matrix

Determine the cross-correlations of atomic displacements.

dccm(x, ...)

Arguments

x
a numeric matrix of Cartesian coordinates with a row per structure/frame which will br passed to dccm.xyz(). Alternatively, an object of class nma as obtained from function nma that will be passed to the dccm.nma() function, see below for examples.

...
additional arguments passed to the methods dccm.xyz, dccm.pca, dccm.nma, and dccm.enma.

Details

dccm is a generic function calling the corresponding function determined by the class of the input argument x. Use methods("dccm") to get all the methods for dccm generic:

dccm.xyz will be used when x is a numeric matrix containing Cartesian coordinates (e.g. trajectory data).

dccm.pca will calculate the cross-correlations based on an pca object.

dccm.nma will calculate the cross-correlations based on an nma object. Similarly, dccm.enma will calculate the correlation matrices based on an ensemble of nma objects (as obtained from function nma.pdbs).

plot.dccm and pymol.dccm provides convenient functionality to plot a correlation map, and visualize the correlations in the structure, respectively.

See examples for each corresponding function for more details.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Barry Grant, Lars Skjaerven

See also