Dynamic Cross-Correlation from Normal Modes Analysis

Calculate the cross-correlation matrix from Normal Modes Analysis.

# S3 method for nma
dccm(x, nmodes = NULL, ncore = NULL, progress = NULL, ...)

Arguments

x
an object of class nma as obtained from function nma.

nmodes
numerical, number of modes to consider.

ncore
number of CPU cores used to do the calculation. ncore>1 requires package ‘parallel’ installed.

progress
progress bar for use with shiny web app.

...
additional arguments ?

Details

This function calculates the cross-correlation matrix from Normal Modes Analysis (NMA) obtained from nma of a protein structure. It returns a matrix of residue-wise cross-correlations whose elements, Cij, may be displayed in a graphical representation frequently termed a dynamical cross-correlation map, or DCCM.

If Cij = 1 the fluctuations of residues i and j are completely correlated (same period and same phase), if Cij = -1 the fluctuations of residues i and j are completely anticorrelated (same period and opposite phase), and if Cij = 0 the fluctuations of i and j are not correlated.

Value

Returns a cross-correlation matrix.

References

Wynsberghe. A.W.V, Cui, Q. Structure 14, 1647--1653. Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Lars Skjaerven

See also

Examples

## Fetch stucture
pdb <- read.pdb( system.file("examples/1hel.pdb", package="bio3d") )

## Calculate normal modes
modes <- nma(pdb)

## Calculate correlation matrix
cm <- dccm.nma(modes)

## Plot correlation map
plot(cm, sse = pdb, contour = FALSE, col.regions = bwr.colors(20),
     at = seq(-1, 1, 0.1))