Periodic Table of the Elements

This data set gives various information on chemical elements.

elements

Format

A data frame containing for each chemical element the following information.

num

atomic number

symb

elemental symbol

areneg

Allred and Rochow electronegativity (0.0 if unknown)

rcov

covalent radii (in Angstrom) (1.6 if unknown)

rbo

"bond order" radii

rvdw

van der Waals radii (in Angstrom) (2.0 if unknown)

maxbnd

maximum bond valence (6 if unknown)

mass

IUPAC recommended atomic masses (in amu)

elneg

Pauling electronegativity (0.0 if unknown)

ionization

ionization potential (in eV) (0.0 if unknown)

elaffinity

electron affinity (in eV) (0.0 if unknown)

red

red value for visualization

green

green value for visualization

blue

blue value for visualization

name

element name

Source

Open Babel (2.3.1) file: element.txt

Created from the Blue Obelisk Cheminformatics Data Repository
Direct Source: http://www.blueobelisk.org/
http://www.blueobelisk.org/repos/blueobelisk/elements.xml includes furhter bibliographic citation information

- Allred and Rochow Electronegativity from http://www.hull.ac.uk/chemistry/electroneg.php?type=Allred-Rochow
- Covalent radii from http://dx.doi.org/10.1039/b801115j
- Van der Waals radii from http://dx.doi.org/10.1021/jp8111556

Examples

data(elements)
elements

# Get the mass of some elements
symb <- c("C","O","H")
elements[match(symb,elements[,"symb"]),"mass"]

# Get the van der Waals radii of some elements
symb <- c("C","O","H")
elements[match(symb,elements[,"symb"]),"rvdw"]