Bio3d Example Data

These data sets contain the results of running various Bio3D functions on example kinesin and transducin structural data, and on a short coarse-grained MD simulation data for HIV protease. The main purpose of including this data (which may be generated by the user by following the extended examples documented within the various Bio3D functions) is to speed up example execution. It should allow users to more quickly appreciate the capabilities of functions that would otherwise require raw data download, input and processing before execution.

Note that related datasets formed the basis of the work described in (Grant, 2007) and (Yao & Grant, 2013) for kinesin and transducin examples, respectively.

data(kinesin)
data(transducin)
data(hivp)

Format

Three objects from analysis of the kinesin and transducin sequence and structure data:

pdbs is a list of class pdbs containing aligned PDB structure data. In the case of transducin this is the output of running pdbaln on a set of 53 G[alpha]i structures from the PDB database (see pdbs$id or annotation described below for details). The coordinates are fitted onto the first structure based on "core" positions obtained from core.find and superposed using the function pdbfit.

core is a list of class "core" obtained by running the function core.find on the pdbs object as described above.

annotation is a character matrix describing the nucleotide state and bound ligand species for each structure in pdbs as obtained from the function pdb.annotate.

One object named net in the hivp example data stores the correlation network obtained from the analysis of the MD simulation trajectory of HIV protease using the cna function. The original trajectory file can be accessed by the command ‘system.file("examples/hivp.dcd", package="bio3d")’.

Source

A related but more extensive dataset formed the basis of the work described in (Grant, 2007) and (Yao & Grant, 2013) for kinesin and transducin examples, respectively.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Grant, B.J. et al. (2007) J. Mol. Biol. 368, 1231--1248.

Yao, X.Q. et al. (2013) Biophys. J. 105, L08--L10.