NMA Fluctuations

Calculates the atomic fluctuations from normal modes analysis.

fluct.nma(nma, mode.inds=NULL)

Arguments

nma
a list object of class "nma" (obtained with nma).

mode.inds
a numeric vector containing the the mode numbers in which the calculation should be based.

Details

Atomic fluctuations are calculated based on the nma object. By default all modes are included in the calculation.

See examples for more details.

Value

Returns a numeric vector of atomic fluctuations.

References

Hinsen, K. et al. (2000) Chemical Physics 261, 25--37. Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Lars Skjaerven

See also

Examples

## Fetch stucture
pdb <- read.pdb( system.file("examples/1hel.pdb", package="bio3d") )

## Calculate (vibrational) normal modes
modes <- nma(pdb)

#>  Building Hessian...		Done in 0.012 seconds.
#>  Diagonalizing Hessian...	Done in 0.084 seconds.

## Fluctuations
f <- fluct.nma(modes)

## Fluctuations of first non-trivial mode
f <- fluct.nma(modes, mode.inds=c(7,8))