Is an Object of Class ‘pdb(s)’?

Checks whether its argument is an object of class ‘pdb’ or ‘pdbs’.

is.pdb(x)
  is.pdbs(x)

Arguments

x
an R object.

Details

Tests if the object ‘x’ is of class ‘pdb’ (is.pdb) or ‘pdbs’ (is.pdbs), i.e. if ‘x’ has a “class” attribute equal to pdb or pdbs.

Value

TRUE if x is an object of class ‘pdb(s)’ and FALSE otherwise

See also

Examples

# Read a PDB file
pdb <- read.pdb( system.file("examples/1hel.pdb", package="bio3d") )
is.pdb(pdb)

#> [1] TRUE