PCA / NMA Atomic Displacement Trajectory

Make a trajectory of atomic displacments along a given principal component / normal mode.

mktrj(...)

# S3 method for pca
mktrj(pca = NULL, pc = 1, mag = 1, step = 0.125, file =
NULL, pdb = NULL, rock=TRUE, ...)

# S3 method for nma
mktrj(nma = NULL, mode = 7, mag = 10, step = 1.25, file = NULL, 
      pdb = NULL, rock=TRUE,  ...)

# S3 method for enma
mktrj(enma = NULL, pdbs = NULL, s.inds = NULL, m.inds = NULL,
      mag = 10, step = 1.25, file = NULL, rock = TRUE, ncore = NULL, ...)

Arguments

pca
an object of class "pca" as obtained with function pca.xyz or pca.

nma
an object of class "nma" as obtained with function nma.pdb.

enma
an object of class "enma" as obtained with function nma.pdbs.

pc
the PC number along which displacements should be made.

mag
a magnification factor for scaling the displacements.

step
the step size by which to increment along the pc/mode.

file
a character vector giving the output PDB file name.

pdb
an object of class "pdb" as obtained from read.pdb or class "pdbs" as obtained from read.fasta.pdb. If not NULL, used as reference to write the PDB file.

rock
logical, if TRUE the trajectory rocks.

mode
the mode number along which displacements should be made.

pdbs
a list object of class "pdbs" (obtained with pdbaln or read.fasta.pdb) which corresponds to the "enma" object.

s.inds
index or indices pointing to the structure(s) in the enma object for which the trajectory shall be generated.

m.inds
the mode number(s) along which displacements should be made.

ncore
number of CPU cores used to do the calculation. ncore>1 requires package ‘parallel’ installed.

...
additional arguments passed to and from functions (e.g. to function write.pdb).

Details

Trajectory frames are built from reconstructed Cartesian coordinates produced by interpolating from the mean structure along a given pc or mode, in increments of step.

An optional magnification factor can be used to amplify displacements. This involves scaling by mag-times the standard deviation of the conformer distribution along the given pc (i.e. the square root of the associated eigenvalue).

Note

Molecular graphics software such as VMD or PyMOL is useful for viewing trajectories see e.g:
http://www.ks.uiuc.edu/Research/vmd/.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Barry Grant, Lars Skjaerven

See also

Examples

if (FALSE) {

##- PCA example
attach(transducin)

# Calculate principal components
pc.xray <- pca(pdbs, fit=TRUE)

# Write PC trajectory of pc=1
outfile = tempfile()
a <- mktrj(pc.xray, file = outfile)
outfile

detach(transducin)

##- NMA example
## Fetch stucture
pdb <- read.pdb( system.file("examples/1hel.pdb", package="bio3d") )

## Calculate (vibrational) normal modes
modes <- nma(pdb)

## Visualize modes
outfile = file.path(tempdir(), "mode_7.pdb")
mktrj(modes, mode=7, pdb=pdb, file = outfile)
outfile

}