Pair Indices

A utility function to determine indices for pairwise comparisons.

pairwise(N)

Arguments

N
a single numeric value representing the total number of things to undergo pairwise comparison.

Value

Returns a two column numeric matrix giving the indices for all pairs.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Barry Grant

See also

Examples

  pairwise(3)

#>      [,1] [,2]
#> [1,]    1    2
#> [2,]    1    3
#> [3,]    2    3

  pairwise(20)

#>        [,1] [,2]
#>   [1,]    1    2
#>   [2,]    1    3
#>   [3,]    1    4
#>   [4,]    1    5
#>   [5,]    1    6
#>   [6,]    1    7
#>   [7,]    1    8
#>   [8,]    1    9
#>   [9,]    1   10
#>  [10,]    1   11
#>  [11,]    1   12
#>  [12,]    1   13
#>  [13,]    1   14
#>  [14,]    1   15
#>  [15,]    1   16
#>  [16,]    1   17
#>  [17,]    1   18
#>  [18,]    1   19
#>  [19,]    1   20
#>  [20,]    2    3
#>  [21,]    2    4
#>  [22,]    2    5
#>  [23,]    2    6
#>  [24,]    2    7
#>  [25,]    2    8
#>  [26,]    2    9
#>  [27,]    2   10
#>  [28,]    2   11
#>  [29,]    2   12
#>  [30,]    2   13
#>  [31,]    2   14
#>  [32,]    2   15
#>  [33,]    2   16
#>  [34,]    2   17
#>  [35,]    2   18
#>  [36,]    2   19
#>  [37,]    2   20
#>  [38,]    3    4
#>  [39,]    3    5
#>  [40,]    3    6
#>  [41,]    3    7
#>  [42,]    3    8
#>  [43,]    3    9
#>  [44,]    3   10
#>  [45,]    3   11
#>  [46,]    3   12
#>  [47,]    3   13
#>  [48,]    3   14
#>  [49,]    3   15
#>  [50,]    3   16
#>  [51,]    3   17
#>  [52,]    3   18
#>  [53,]    3   19
#>  [54,]    3   20
#>  [55,]    4    5
#>  [56,]    4    6
#>  [57,]    4    7
#>  [58,]    4    8
#>  [59,]    4    9
#>  [60,]    4   10
#>  [61,]    4   11
#>  [62,]    4   12
#>  [63,]    4   13
#>  [64,]    4   14
#>  [65,]    4   15
#>  [66,]    4   16
#>  [67,]    4   17
#>  [68,]    4   18
#>  [69,]    4   19
#>  [70,]    4   20
#>  [71,]    5    6
#>  [72,]    5    7
#>  [73,]    5    8
#>  [74,]    5    9
#>  [75,]    5   10
#>  [76,]    5   11
#>  [77,]    5   12
#>  [78,]    5   13
#>  [79,]    5   14
#>  [80,]    5   15
#>  [81,]    5   16
#>  [82,]    5   17
#>  [83,]    5   18
#>  [84,]    5   19
#>  [85,]    5   20
#>  [86,]    6    7
#>  [87,]    6    8
#>  [88,]    6    9
#>  [89,]    6   10
#>  [90,]    6   11
#>  [91,]    6   12
#>  [92,]    6   13
#>  [93,]    6   14
#>  [94,]    6   15
#>  [95,]    6   16
#>  [96,]    6   17
#>  [97,]    6   18
#>  [98,]    6   19
#>  [99,]    6   20
#> [100,]    7    8
#> [101,]    7    9
#> [102,]    7   10
#> [103,]    7   11
#> [104,]    7   12
#> [105,]    7   13
#> [106,]    7   14
#> [107,]    7   15
#> [108,]    7   16
#> [109,]    7   17
#> [110,]    7   18
#> [111,]    7   19
#> [112,]    7   20
#> [113,]    8    9
#> [114,]    8   10
#> [115,]    8   11
#> [116,]    8   12
#> [117,]    8   13
#> [118,]    8   14
#> [119,]    8   15
#> [120,]    8   16
#> [121,]    8   17
#> [122,]    8   18
#> [123,]    8   19
#> [124,]    8   20
#> [125,]    9   10
#> [126,]    9   11
#> [127,]    9   12
#> [128,]    9   13
#> [129,]    9   14
#> [130,]    9   15
#> [131,]    9   16
#> [132,]    9   17
#> [133,]    9   18
#> [134,]    9   19
#> [135,]    9   20
#> [136,]   10   11
#> [137,]   10   12
#> [138,]   10   13
#> [139,]   10   14
#> [140,]   10   15
#> [141,]   10   16
#> [142,]   10   17
#> [143,]   10   18
#> [144,]   10   19
#> [145,]   10   20
#> [146,]   11   12
#> [147,]   11   13
#> [148,]   11   14
#> [149,]   11   15
#> [150,]   11   16
#> [151,]   11   17
#> [152,]   11   18
#> [153,]   11   19
#> [154,]   11   20
#> [155,]   12   13
#> [156,]   12   14
#> [157,]   12   15
#> [158,]   12   16
#> [159,]   12   17
#> [160,]   12   18
#> [161,]   12   19
#> [162,]   12   20
#> [163,]   13   14
#> [164,]   13   15
#> [165,]   13   16
#> [166,]   13   17
#> [167,]   13   18
#> [168,]   13   19
#> [169,]   13   20
#> [170,]   14   15
#> [171,]   14   16
#> [172,]   14   17
#> [173,]   14   18
#> [174,]   14   19
#> [175,]   14   20
#> [176,]   15   16
#> [177,]   15   17
#> [178,]   15   18
#> [179,]   15   19
#> [180,]   15   20
#> [181,]   16   17
#> [182,]   16   18
#> [183,]   16   19
#> [184,]   16   20
#> [185,]   17   18
#> [186,]   17   19
#> [187,]   17   20
#> [188,]   18   19
#> [189,]   18   20
#> [190,]   19   20