Plot eNMA Results

Produces a plot of atomic fluctuations of aligned normal modes.

# S3 method for enma
plot(x,
        pdbs = NULL,
        xlab = NULL,
        ylab="Fluctuations", ...)

Arguments

x
the results of ensemble NMA obtained with nma.pdbs. Alternatively, a matrix in the similar format as enma$fluctuations can be provided.

pdbs
an object of class ‘pdbs’ in which the ‘enma’ object x was obtained from. If provided SSE data of the first structure of pdbs will drawn.

xlab
a label for the x axis.

ylab
labels for the y axes.

...
extra plotting arguments passed to plot.fluct that effect the atomic fluctuations plot only.

Details

plot.enma produces a fluctuation plot of aligned nma objects. If corresponding pdbs object is provided the plot contains SSE annotation and appropriate resiude index numbering.

Value

Called for its effect.

References

Skjaerven, L. et al. (2014) BMC Bioinformatics 15, 399. Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Lars Skjaerven, Barry Grant

See also

Examples

if (FALSE) {
ids <- c("1a70_A", "1czp_A", "1frd_A", "1fxi_A", "1iue_A", "1pfd_A")
raw.files <- get.pdb(ids, path = "raw_pdbs")
files <- pdbsplit(raw.files, ids, path = "raw_pdbs/split_chain")

## Sequence/structure alignement
pdbs <- pdbaln(files)

## Normal mode analysis on aligned data
modes <- nma(pdbs)

## Plot fluctuations
plot(modes, pdbs=pdbs)

## Group and spread fluctuation profiles
hc <- hclust(as.dist(1-modes$rmsip))
col <- cutree(hc, k=2)
plot(modes, pdbs=pdbs, col=col, spread=TRUE)
}