Plot Geostas Results

Plot an atomic movement similarity matrix with domain annotation

# S3 method for geostas
plot(x, at=seq(0, 1, 0.1), main="AMSM with Domain Assignment",
                         col.regions=rev(heat.colors(200)), 
                         margin.segments=x$grps, ...)

Arguments

x
an object of type geostas as obtained by the ‘geostas’ function.

at
numeric vector specifying the levels to be colored.

main
a main title for the plot.

col.regions
color vector. See contourplot for more information.

margin.segments
a numeric vector of cluster membership as obtained from cutree() or other community detection method. This will be used for bottom and left margin annotation.

...
additional graphical parameters for plot.dccm and contourplot.

Details

This is a wrapper function for plot.dccm with appropriate adjustments for plotting atomic movement similarity matrix obtained from function geostas.

See the plot.dccm for more details.

Value

Called for its effect.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Barry Grant, Lars Skjaerven

See also