Read CRD File

Read a CHARMM CARD (CRD) coordinate file.

# S3 method for charmm
read.crd(file, ext = TRUE, verbose = TRUE, ...)

Arguments

file
the name of the CRD file to be read.

ext
logical, if TRUE assume expanded CRD format.

verbose
print details of the reading process.

...
arguments going nowhere.

Details

See the function read.pdb for more details.

Value

Returns a list with the following components:

atom

a character matrix containing all atomic coordinate data, with a row per atom and a column per record type. See below for details of the record type naming convention (useful for accessing columns).

xyz

a numeric vector of coordinate data.

calpha

logical vector with length equal to nrow(atom) with TRUE values indicating a C-alpha “elety”.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

For a description of CHARMM CARD (CRD) format see:
http://www.charmmtutorial.org/index.php/CHARMM:The_Basics.

Author

Barry Grant

Note

Similar to the output of read.pdb, the column names of atom can be used as a convenient means of data access, namely: Atom serial number “eleno”, Atom type “elety”, Alternate location indicator “alt”, Residue name “resid”, Residue sequence number “resno”, Code for insertion of residues “insert”, Orthogonal coordinates “x”, Orthogonal coordinates “y”, Orthogonal coordinates “z”, Weighting factor “b”. See examples for further details.

See also

Examples

if (FALSE) {
 pdb <- read.pdb("1bg2")
 crdfile <- paste(tempfile(), '.crd', sep='')
 write.crd(pdb, file=crdfile)
 crd <- read.crd(crdfile, ext=FALSE)
 ca.inds <- which(crd$calpha)
 crd$atom[ca.inds[1:20],c("x","y","z")]
# write.pdb(crd, file=tempfile())
}