Filter or Trim a PDBs Object

Trim residues and/or filter out structures from a PDBs object.

# S3 method for pdbs
trim(pdbs, row.inds=NULL, col.inds=NULL, ...)

Arguments

pdbs
an object of class pdbs as obtained from function pdbaln or read.fasta.pdb; a xyz matrix containing the cartesian coordinates of C-alpha atoms.

row.inds
a numeric vector of indices pointing to the PDB structures to keep (rows in the pdbs$ali matrix).

col.inds
a numeric vector of indices pointing to the alignment columns to keep (columns in the pdbs$ali matrix).

...
additional arguments passed to and from functions.

Details

Utility function to remove structures, or trim off columns, in a ‘pdbs’ object.

Value

Returns an updated ‘pdbs’ object with the following components:

xyz

numeric matrix of aligned C-alpha coordinates.

resno

character matrix of aligned residue numbers.

b

numeric matrix of aligned B-factor values.

chain

character matrix of aligned chain identifiers.

id

character vector of PDB sequence/structure names.

ali

character matrix of aligned sequences.

call

the matched call.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Lars Skjaerven

See also

Examples

if (FALSE) {
## Fetch PDB files and split to chain A only PDB files
ids <- c("1a70_A", "1czp_A", "1frd_A", "1fxi_A", "1iue_A", "1pfd_A")
raw.files <- get.pdb(ids, path = "raw_pdbs")
files <- pdbsplit(raw.files, ids, path = "raw_pdbs/split_chain")

## Sequence Alignement, and connectivity check
pdbs <- pdbaln(files)

cons <- inspect.connectivity(pdbs)

## omit files with missing residues
trim.pdbs(pdbs, row.inds=which(cons))
}