Fetch UniProt Entry Data.

Fetch protein sequence and functional information from the UniProt database.

uniprot(accid)

Arguments

accid
UniProt accession id.

Details

This is a basic utility function for downloading information from the UniProt database. UniProt contains protein sequence and functional information.

Value

Returns a list object with the following components:

accession

a character vector with UniProt accession id's.

name

abbreviated name.

fullName

full recommended protein name.

shortName

short protein name.

sequence

protein sequence.

gene

gene names.

organism

organism.

taxon

taxonomic lineage.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

See also the UniProt web-site for more information:
http://www.uniprot.org/.

Author

Lars Skjaerven

See also

Examples

if (FALSE) {
# UNIPROT server connection required - testing excluded

prot <- uniprot('PH4H_HUMAN')
prot$fullName
prot$sequence
}