Write CRD File

Write a CHARMM CARD (CRD) coordinate file.

write.crd(pdb = NULL, xyz = pdb$xyz, resno = NULL, resid = NULL, 
   eleno = NULL, elety = NULL, segid = NULL, resno2 = NULL, b = NULL, 
   verbose = FALSE, file = "R.crd")

Arguments

pdb
a structure object obtained from read.pdb or read.crd.

xyz
Cartesian coordinates as a vector or 3xN matrix.

resno
vector of residue numbers of length equal to length(xyz)/3.

resid
vector of residue types/ids of length equal to length(xyz)/3.

eleno
vector of element/atom numbers of length equal to length(xyz)/3.

elety
vector of element/atom types of length equal to length(xyz)/3.

segid
vector of segment identifiers with length equal to length(xyz)/3.

resno2
vector of alternate residue numbers of length equal to length(xyz)/3.

b
vector of weighting factors of length equal to length(xyz)/3.

verbose
logical, if TRUE progress details are printed.

file
the output file name.

Details

Only the xyz argument is strictly required. Other arguments assume a default poly-ALA C-alpha structure with a blank segid and B-factors equal to 0.00.

Value

Called for its effect.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

For a description of CHARMM CARD (CRD) format see:
http://www.charmmtutorial.org/index.php/CHARMM:The_Basics.

Author

Barry Grant

Note

Check that resno and eleno do not exceed “9999”.

See also

Examples

if (FALSE) {
# Read a PDB file
pdb <- read.pdb( "1bg2" )
summary(pdb)
# Convert to CHARMM format
new <- convert.pdb(pdb, type="charmm")
summary(new)
# Write a CRD file
write.crd(new, file="4charmm.crd")
}