AAindex: Amino Acid Index Database

A collection of published indices, or scales, of numerous physicochemical and biological properties of the 20 standard aminoacids (Release 9.1, August 2006).

data(aa.index)

Format

A list of 544 named indeces each with the following components:

H character vector: Accession number.

D character vector: Data description.

R character vector: LITDB entry number.

A character vector: Author(s).

T character vector: Title of the article.

J character vector: Journal reference.

C named numeric vector: Correlation coefficients of similar indeces (with coefficients of 0.8/-0.8 or more/less). The correlation coefficient is calculated with zeros filled for missing values.

I named numeric vector: Amino acid index data.

Source

‘AAIndex’ was obtained from:
http://www.genome.jp/aaindex/
For a description of the ‘AAindex’ database see:
http://www.genome.jp/aaindex/aaindex_help.html.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

‘AAIndex’ is the work of Kanehisa and co-workers:
Kawashima and Kanehisa (2000) Nucleic Acids Res. 28, 374;
Tomii and Kanehisa (1996) Protein Eng. 9, 27--36;
Nakai, Kidera and Kanehisa (1988) Protein Eng. 2, 93--100.

Examples

## Load AAindex data
data(aa.index)

## Find all indeces described as "volume"
ind <- which(sapply(aa.index, function(x)
                    length(grep("volume", x$D, ignore.case=TRUE)) != 0))

## find all indeces with author "Kyte"
ind <- which(sapply(aa.index, function(x) length(grep("Kyte", x$A)) != 0))

## examine the index
aa.index[[ind]]$I

#>    A    R    N    D    C    Q    E    G    H    I    L    K    M    F    P    S 
#>  1.8 -4.5 -3.5 -3.5  2.5 -3.5 -3.5 -0.4 -3.2  4.5  3.8 -3.9  1.9  2.8 -1.6 -0.8 
#>    T    W    Y    V 
#> -0.7 -0.9 -1.3  4.2 

## find indeces which correlate with it
all.ind <- names(which(Mod(aa.index[[ind]]$C) >= 0.88))

## examine them all
sapply(all.ind, function (x) aa.index[[x]]$I)

#>   JURD980101 CHOC760103 OLSK800101 JANJ780102 NADH010102 NADH010101 DESM900102
#> A       1.10       0.38       1.38         51         51         58       1.26
#> R      -5.10       0.01       0.00          5       -144       -184       0.38
#> N      -3.50       0.12       0.37         22        -84        -93       0.59
#> D      -3.60       0.15       0.52         19        -78        -97       0.27
#> C       2.50       0.45       1.43         74        137        116       1.60
#> Q      -3.68       0.07       0.22         16       -128       -139       0.39
#> E      -3.20       0.18       0.71         16       -115       -131       0.23
#> G      -0.64       0.36       1.34         52        -13        -11       1.08
#> H      -3.20       0.17       0.66         34        -55        -73       1.00
#> I       4.50       0.60       2.32         66        106        107       1.44
#> L       3.80       0.45       1.47         60        103         95       1.36
#> K      -4.11       0.03       0.15          3       -205        -24       0.33
#> M       1.90       0.40       1.78         52         73         78       1.52
#> F       2.80       0.50       1.72         58        108         92       1.46
#> P      -1.90       0.18       0.85         25        -79        -79       0.54
#> S      -0.50       0.22       0.86         35        -26        -34       0.98
#> T      -0.70       0.23       0.89         30         -3         -7       1.01
#> W      -0.46       0.27       0.82         49         69         59       1.06
#> Y      -1.30       0.15       0.47         24         11        -11       0.89
#> V       4.20       0.54       1.99         64        108        100       1.33
#>   EISD860103 CHOC760104 NADH010103 WOLR810101 RADA880101 MANP780101 KUHL950101
#> A       0.00       0.20         41       1.94       1.81      12.97       0.78
#> R      -0.96       0.00       -109     -19.92     -14.92      11.72       1.58
#> N      -0.86       0.03        -74      -9.68      -6.64      11.42       1.20
#> D      -0.98       0.04        -47     -10.95      -8.72      10.85       1.35
#> C       0.76       0.22        169      -1.24       1.28      14.63       0.55
#> Q      -1.00       0.01       -104      -9.38      -5.54      11.76       1.19
#> E      -0.89       0.03        -90     -10.20      -6.81      11.89       1.45
#> G       0.00       0.18        -18       2.39       0.94      12.43       0.68
#> H      -0.75       0.02        -35     -10.27      -4.66      12.16       0.99
#> I       0.99       0.19        104       2.15       4.92      15.67       0.47
#> L       0.89       0.16        103       2.28       4.92      14.90       0.56
#> K      -0.99       0.00       -148      -9.52      -5.55      11.36       1.10
#> M       0.94       0.11         77      -1.48       2.35      14.39       0.66
#> F       0.92       0.14        128      -0.76       2.98      14.00       0.47
#> P       0.22       0.04        -81      -3.68       0.00      11.37       0.69
#> S      -0.67       0.08        -31      -5.06      -3.40      11.23       1.00
#> T       0.09       0.08         10      -4.88      -2.57      11.69       1.05
#> W       0.67       0.04        102      -5.88       2.33      13.93       0.70
#> Y      -0.93       0.03         36      -6.11      -0.14      13.42       1.00
#> V       0.84       0.18        116       1.99       4.04      15.71       0.51
#>   GUYH850105 OOBM770101
#> A      -0.27     -1.895
#> R       2.00     -1.475
#> N       0.61     -1.560
#> D       0.50     -1.518
#> C      -0.23     -2.035
#> Q       1.00     -1.521
#> E       0.33     -1.535
#> G      -0.22     -1.898
#> H       0.37     -1.755
#> I      -0.80     -1.951
#> L      -0.44     -1.966
#> K       1.17     -1.374
#> M      -0.31     -1.963
#> F      -0.55     -1.864
#> P       0.36     -1.699
#> S       0.17     -1.753
#> T       0.18     -1.767
#> W       0.05     -1.869
#> Y       0.48     -1.686
#> V      -0.65     -1.981