Periodic Table of the Elements

This data set gives various information on chemical elements.

elements

Format

A data frame containing for each chemical element the following information.

num

atomic number

symb

elemental symbol

areneg

Allred and Rochow electronegativity (0.0 if unknown)

rcov

covalent radii (in Angstrom) (1.6 if unknown)

rbo

"bond order" radii

rvdw

van der Waals radii (in Angstrom) (2.0 if unknown)

maxbnd

maximum bond valence (6 if unknown)

mass

IUPAC recommended atomic masses (in amu)

elneg

Pauling electronegativity (0.0 if unknown)

ionization

ionization potential (in eV) (0.0 if unknown)

elaffinity

electron affinity (in eV) (0.0 if unknown)

red

red value for visualization

green

green value for visualization

blue

blue value for visualization

name

element name

Source

Open Babel (2.3.1) file: element.txt

Created from the Blue Obelisk Cheminformatics Data Repository
Direct Source: http://www.blueobelisk.org/
http://www.blueobelisk.org/repos/blueobelisk/elements.xml includes furhter bibliographic citation information

- Allred and Rochow Electronegativity from http://www.hull.ac.uk/chemistry/electroneg.php?type=Allred-Rochow
- Covalent radii from http://dx.doi.org/10.1039/b801115j
- Van der Waals radii from http://dx.doi.org/10.1021/jp8111556

Examples

data(elements)
elements

#>     num symb areneg rcov  rbo rvdw maxbnd       mass elneg ionization
#> 1     0   Xx   0.00 0.00 0.00 0.00      0   0.000000  0.00     0.0000
#> 2     1    H   2.20 0.31 0.31 1.10      1   1.007940  2.20    13.5984
#> 3     2   He   0.00 0.28 0.28 1.40      0   4.002602  0.00    24.5874
#> 4     3   Li   0.97 1.28 1.28 1.81      1   6.941000  0.98     5.3917
#> 5     4   Be   1.47 0.96 0.96 1.53      2   9.012182  1.57     9.3227
#> 6     5    B   2.01 0.84 0.84 1.92      4  10.811000  2.04     8.2980
#> 7     6    C   2.50 0.76 0.76 1.70      4  12.010700  2.55    11.2603
#> 8     7    N   3.07 0.71 0.71 1.55      4  14.006700  3.04    14.5341
#> 9     8    O   3.50 0.66 0.66 1.52      2  15.999400  3.44    13.6181
#> 10    9    F   4.10 0.57 0.57 1.47      1  18.998403  3.98    17.4228
#> 11   10   Ne   0.00 0.58 0.58 1.54      0  20.179700  0.00    21.5645
#> 12   11   Na   1.01 1.66 1.66 2.27      1  22.989770  0.93     5.1391
#> 13   12   Mg   1.23 1.41 1.41 1.73      2  24.305000  1.31     7.6462
#> 14   13   Al   1.47 1.21 1.21 1.84      6  26.981538  1.61     5.9858
#> 15   14   Si   1.74 1.11 1.11 2.10      6  28.085500  1.90     8.1517
#> 16   15    P   2.06 1.07 1.07 1.80      6  30.973761  2.19    10.4867
#> 17   16    S   2.44 1.05 1.05 1.80      6  32.065000  2.58    10.3600
#> 18   17   Cl   2.83 1.02 1.02 1.75      1  35.453000  3.16    12.9676
#> 19   18   Ar   0.00 1.06 1.06 1.88      0  39.948000  0.00    15.7596
#> 20   19    K   0.91 2.03 2.03 2.75      1  39.098300  0.82     4.3407
#> 21   20   Ca   1.04 1.76 1.76 2.31      2  40.078000  1.00     6.1132
#> 22   21   Sc   1.20 1.70 1.70 2.30      6  44.955910  1.36     6.5615
#> 23   22   Ti   1.32 1.60 1.60 2.15      6  47.867000  1.54     6.8281
#> 24   23    V   1.45 1.53 1.53 2.05      6  50.941500  1.63     6.7462
#> 25   24   Cr   1.56 1.39 1.39 2.05      6  51.996100  1.66     6.7665
#> 26   25   Mn   1.60 1.39 1.39 2.05      8  54.938049  1.55     7.4340
#> 27   26   Fe   1.64 1.32 1.32 2.05      6  55.845000  1.83     7.9024
#> 28   27   Co   1.70 1.26 1.26 2.00      6  58.933200  1.88     7.8810
#> 29   28   Ni   1.75 1.24 1.24 2.00      6  58.693400  1.91     7.6398
#> 30   29   Cu   1.75 1.32 1.32 2.00      6  63.546000  1.90     7.7264
#> 31   30   Zn   1.66 1.22 1.22 2.10      6  65.380000  1.65     9.3942
#> 32   31   Ga   1.82 1.22 1.22 1.87      3  69.723000  1.81     5.9993
#> 33   32   Ge   2.02 1.20 1.20 2.11      4  72.640000  2.01     7.8994
#> 34   33   As   2.20 1.19 1.19 1.85      3  74.921600  2.18     9.7886
#> 35   34   Se   2.48 1.20 1.20 1.90      2  78.960000  2.55     9.7524
#> 36   35   Br   2.74 1.20 1.20 1.83      1  79.904000  2.96    11.8138
#> 37   36   Kr   0.00 1.16 1.16 2.02      0  83.798000  3.00    13.9996
#> 38   37   Rb   0.89 2.20 2.20 3.03      1  85.467800  0.82     4.1771
#> 39   38   Sr   0.99 1.95 1.95 2.49      2  87.620000  0.95     5.6949
#> 40   39    Y   1.11 1.90 1.90 2.40      6  88.905850  1.22     6.2173
#> 41   40   Zr   1.22 1.75 1.75 2.30      6  91.224000  1.33     6.6339
#> 42   41   Nb   1.23 1.64 1.64 2.15      6  92.906380  1.60     6.7589
#> 43   42   Mo   1.30 1.54 1.54 2.10      6  95.960000  2.16     7.0924
#> 44   43   Tc   1.36 1.47 1.47 2.05      6  98.000000  1.90     7.2800
#> 45   44   Ru   1.42 1.46 1.46 2.05      6 101.070000  2.20     7.3605
#> 46   45   Rh   1.45 1.42 1.42 2.00      6 102.905500  2.28     7.4589
#> 47   46   Pd   1.35 1.39 1.39 2.05      6 106.420000  2.20     8.3369
#> 48   47   Ag   1.42 1.45 1.45 2.10      6 107.868200  1.93     7.5762
#> 49   48   Cd   1.46 1.44 1.44 2.20      6 112.411000  1.69     8.9938
#> 50   49   In   1.49 1.42 1.42 2.20      3 114.818000  1.78     5.7864
#> 51   50   Sn   1.72 1.39 1.39 1.93      4 118.701000  1.96     7.3439
#> 52   51   Sb   1.82 1.39 1.39 2.17      3 121.760000  2.05     8.6084
#> 53   52   Te   2.01 1.38 1.38 2.06      2 127.600000  2.10     9.0096
#> 54   53    I   2.21 1.39 1.39 1.98      1 126.904470  2.66    10.4513
#> 55   54   Xe   0.00 1.40 1.40 2.16      0 131.293000  2.60    12.1298
#> 56   55   Cs   0.86 2.44 2.44 3.43      1 132.905450  0.79     3.8939
#> 57   56   Ba   0.97 2.15 2.15 2.68      2 137.327000  0.89     5.2117
#> 58   57   La   1.08 2.07 2.07 2.50     12 138.905500  1.10     5.5769
#> 59   58   Ce   0.00 2.04 2.04 2.48      6 140.116000  1.12     5.5387
#> 60   59   Pr   0.00 2.03 2.03 2.47      6 140.907650  1.13     5.4730
#> 61   60   Nd   0.00 2.01 2.01 2.45      6 144.240000  1.14     5.5250
#> 62   61   Pm   0.00 1.99 1.99 2.43      6 145.000000  0.00     5.5820
#> 63   62   Sm   0.00 1.98 1.98 2.42      6 150.360000  1.17     5.6437
#> 64   63   Eu   0.00 1.98 1.98 2.40      6 151.964000  0.00     5.6704
#> 65   64   Gd   0.00 1.96 1.96 2.38      6 157.250000  1.20     6.1498
#> 66   65   Tb   0.00 1.94 1.94 2.37      6 158.925340  0.00     5.8638
#> 67   66   Dy   0.00 1.92 1.92 2.35      6 162.500000  1.22     5.9389
#> 68   67   Ho   0.00 1.92 1.92 2.33      6 164.930320  1.23     6.0215
#> 69   68   Er   0.00 1.89 1.89 2.32      6 167.259000  1.24     6.1077
#> 70   69   Tm   0.00 1.90 1.90 2.30      6 168.934210  1.25     6.1843
#> 71   70   Yb   0.00 1.87 1.87 2.28      6 173.054000  0.00     6.2542
#> 72   71   Lu   0.00 1.87 1.87 2.27      6 174.966800  1.27     5.4259
#> 73   72   Hf   1.23 1.75 1.75 2.25      6 178.490000  1.30     6.8251
#> 74   73   Ta   1.33 1.70 1.70 2.20      6 180.947900  1.50     7.5496
#> 75   74    W   1.40 1.62 1.62 2.10      6 183.840000  2.36     7.8640
#> 76   75   Re   1.46 1.51 1.51 2.05      6 186.207000  1.90     7.8335
#> 77   76   Os   1.52 1.44 1.44 2.00      6 190.230000  2.20     8.4382
#> 78   77   Ir   1.55 1.41 1.41 2.00      6 192.217000  2.20     8.9670
#> 79   78   Pt   1.44 1.36 1.36 2.05      6 195.078000  2.28     8.9588
#> 80   79   Au   1.42 1.36 1.36 2.10      6 196.966550  2.54     9.2255
#> 81   80   Hg   1.44 1.32 1.32 2.05      6 200.590000  2.00    10.4375
#> 82   81   Tl   1.44 1.45 1.45 1.96      3 204.383300  1.62     6.1082
#> 83   82   Pb   1.55 1.46 1.46 2.02      4 207.200000  2.33     7.4167
#> 84   83   Bi   1.67 1.48 1.48 2.07      3 208.980400  2.02     7.2855
#> 85   84   Po   1.76 1.40 1.40 1.97      2 209.000000  2.00     8.4140
#> 86   85   At   1.90 1.50 1.50 2.02      1 210.000000  2.20     0.0000
#> 87   86   Rn   0.00 1.50 1.50 2.20      0 222.000000  0.00    10.7485
#> 88   87   Fr   0.00 2.60 2.60 3.48      1 223.000000  0.70     4.0727
#> 89   88   Ra   0.00 2.21 2.21 2.83      2 226.000000  0.90     5.2784
#> 90   89   Ac   0.00 2.15 2.15 2.00      6 227.000000  1.10     5.1700
#> 91   90   Th   0.00 2.06 2.06 2.40      6 232.038100  1.30     6.3067
#> 92   91   Pa   0.00 2.00 2.00 2.00      6 231.035880  1.50     5.8900
#> 93   92    U   0.00 1.96 1.96 2.30      6 238.028910  1.38     6.1941
#> 94   93   Np   0.00 1.90 1.90 2.00      6 237.000000  1.36     6.2657
#> 95   94   Pu   0.00 1.87 1.87 2.00      6 244.000000  1.28     6.0260
#> 96   95   Am   0.00 1.80 1.80 2.00      6 243.000000  1.30     5.9738
#> 97   96   Cm   0.00 1.69 1.69 2.00      6 247.000000  1.30     5.9914
#> 98   97   Bk   0.00 1.60 1.60 2.00      6 247.000000  1.30     6.1979
#> 99   98   Cf   0.00 1.60 1.60 2.00      6 251.000000  1.30     6.2817
#> 100  99   Es   0.00 1.60 1.60 2.00      6 252.000000  1.30     6.4200
#> 101 100   Fm   0.00 1.60 1.60 2.00      6 257.000000  1.30     6.5000
#> 102 101   Md   0.00 1.60 1.60 2.00      6 258.000000  1.30     6.5800
#> 103 102   No   0.00 1.60 1.60 2.00      6 259.000000  1.30     6.6500
#> 104 103   Lr   0.00 1.60 1.60 2.00      6 262.000000  0.00     4.9000
#> 105 104   Rf   0.00 1.60 1.60 2.00      6 267.000000  0.00     6.0000
#> 106 105   Db   0.00 1.60 1.60 2.00      6 268.000000  0.00     0.0000
#> 107 106   Sg   0.00 1.60 1.60 2.00      6 271.000000  0.00     0.0000
#> 108 107   Bh   0.00 1.60 1.60 2.00      6 272.000000  0.00     0.0000
#> 109 108   Hs   0.00 1.60 1.60 2.00      6 270.000000  0.00     0.0000
#> 110 109   Mt   0.00 1.60 1.60 2.00      6 268.000000  0.00     0.0000
#> 111 110   Ds   0.00 1.60 1.60 2.00      6 281.000000  0.00     0.0000
#> 112 111   Rg   0.00 1.60 1.60 2.00      6 280.000000  0.00     0.0000
#> 113 112   Cp   0.00 1.60 1.60 2.00      6 285.000000  0.00     0.0000
#> 114 113  Uut   0.00 1.60 1.60 2.00      6 284.000000  0.00     0.0000
#> 115 114  Uuq   0.00 1.60 1.60 2.00      6 289.000000  0.00     0.0000
#> 116 115  Uup   0.00 1.60 1.60 2.00      6 288.000000  0.00     0.0000
#> 117 116  Uuh   0.00 1.60 1.60 2.00      6 293.000000  0.00     0.0000
#>     elaffinity  red green blue          name
#> 1    0.0000000 0.07  0.50 0.70         Dummy
#> 2    0.7542037 1.00  1.00 1.00      Hydrogen
#> 3    0.0000000 0.85  1.00 1.00        Helium
#> 4    0.6180490 0.80  0.50 1.00       Lithium
#> 5    0.0000000 0.76  1.00 0.00     Beryllium
#> 6    0.2797230 1.00  0.71 0.71         Boron
#> 7    1.2621180 0.50  0.50 0.50        Carbon
#> 8   -0.0700000 0.05  0.05 1.00      Nitrogen
#> 9    1.4611120 1.00  0.05 0.05        Oxygen
#> 10   3.4011887 0.70  1.00 1.00      Fluorine
#> 11   0.0000000 0.70  0.89 0.96          Neon
#> 12   0.5479260 0.67  0.36 0.95        Sodium
#> 13   0.0000000 0.54  1.00 0.00     Magnesium
#> 14   0.4328300 0.75  0.65 0.65     Aluminium
#> 15   1.3895210 0.50  0.60 0.60       Silicon
#> 16   0.7465000 1.00  0.50 0.00    Phosphorus
#> 17   2.0771029 1.00  1.00 0.19        Sulfur
#> 18   3.6127240 0.12  0.94 0.12      Chlorine
#> 19   0.0000000 0.50  0.82 0.89         Argon
#> 20   0.5014590 0.56  0.25 0.83     Potassium
#> 21   0.0245500 0.24  1.00 0.00       Calcium
#> 22   0.1880000 0.90  0.90 0.90      Scandium
#> 23   0.0840000 0.75  0.76 0.78      Titanium
#> 24   0.5250000 0.65  0.65 0.67      Vanadium
#> 25   0.6758400 0.54  0.60 0.78      Chromium
#> 26   0.0000000 0.61  0.48 0.78     Manganese
#> 27   0.1510000 0.88  0.40 0.20          Iron
#> 28   0.6633000 0.94  0.56 0.63        Cobalt
#> 29   1.1571600 0.31  0.82 0.31        Nickel
#> 30   1.2357800 0.78  0.50 0.20        Copper
#> 31   0.0000000 0.49  0.50 0.69          Zinc
#> 32   0.4100000 0.76  0.56 0.56       Gallium
#> 33   1.2327120 0.40  0.56 0.56     Germanium
#> 34   0.8140000 0.74  0.50 0.89       Arsenic
#> 35   2.0206700 1.00  0.63 0.00      Selenium
#> 36   3.3635880 0.65  0.16 0.16       Bromine
#> 37   0.0000000 0.36  0.72 0.82       Krypton
#> 38   0.4859160 0.44  0.18 0.69      Rubidium
#> 39   0.0520600 0.00  1.00 0.00     Strontium
#> 40   0.3070000 0.58  1.00 1.00       Yttrium
#> 41   0.4260000 0.58  0.88 0.88     Zirconium
#> 42   0.8930000 0.45  0.76 0.79       Niobium
#> 43   0.7472000 0.33  0.71 0.71    Molybdenum
#> 44   0.5500000 0.23  0.62 0.62    Technetium
#> 45   1.0463800 0.14  0.56 0.56     Ruthenium
#> 46   1.1428900 0.04  0.49 0.55       Rhodium
#> 47   0.5621400 0.00  0.41 0.52     Palladium
#> 48   1.3044700 0.88  0.88 1.00        Silver
#> 49   0.0000000 1.00  0.85 0.56       Cadmium
#> 50   0.4040000 0.65  0.46 0.45        Indium
#> 51   1.1120660 0.40  0.50 0.50           Tin
#> 52   1.0474010 0.62  0.39 0.71      Antimony
#> 53   1.9708750 0.83  0.48 0.00     Tellurium
#> 54   3.0590380 0.58  0.00 0.58        Iodine
#> 55   0.0000000 0.26  0.62 0.69         Xenon
#> 56   0.4716260 0.34  0.09 0.56       Caesium
#> 57   0.1446200 0.00  0.79 0.00        Barium
#> 58   0.4700000 0.44  0.83 1.00     Lanthanum
#> 59   0.5000000 1.00  1.00 0.78        Cerium
#> 60   0.5000000 0.85  1.00 0.78  Praseodymium
#> 61   0.5000000 0.78  1.00 0.78     Neodymium
#> 62   0.5000000 0.64  1.00 0.78    Promethium
#> 63   0.5000000 0.56  1.00 0.78      Samarium
#> 64   0.5000000 0.38  1.00 0.78      Europium
#> 65   0.5000000 0.27  1.00 0.78    Gadolinium
#> 66   0.5000000 0.19  1.00 0.78       Terbium
#> 67   0.5000000 0.12  1.00 0.78    Dysprosium
#> 68   0.5000000 0.00  1.00 0.61       Holmium
#> 69   0.5000000 0.00  0.90 0.46        Erbium
#> 70   0.5000000 0.00  0.83 0.32       Thulium
#> 71   0.5000000 0.00  0.75 0.22     Ytterbium
#> 72   0.5000000 0.00  0.67 0.14      Lutetium
#> 73   0.0000000 0.30  0.76 1.00       Hafnium
#> 74   0.3220000 0.30  0.65 1.00      Tantalum
#> 75   0.8150000 0.13  0.58 0.84      Tungsten
#> 76   0.1500000 0.15  0.49 0.67       Rhenium
#> 77   1.0778000 0.15  0.40 0.59        Osmium
#> 78   1.5643600 0.09  0.33 0.53       Iridium
#> 79   2.1251000 0.96  0.93 0.82      Platinum
#> 80   2.3086100 0.80  0.82 0.12          Gold
#> 81   0.0000000 0.71  0.71 0.76       Mercury
#> 82   0.3770000 0.65  0.33 0.30      Thallium
#> 83   0.3640000 0.34  0.35 0.38          Lead
#> 84   0.9423630 0.62  0.31 0.71       Bismuth
#> 85   1.9000000 0.67  0.36 0.00      Polonium
#> 86   2.8000000 0.46  0.31 0.27      Astatine
#> 87   0.0000000 0.26  0.51 0.59         Radon
#> 88   0.0000000 0.26  0.00 0.40      Francium
#> 89   0.0000000 0.00  0.49 0.00        Radium
#> 90   0.0000000 0.44  0.67 0.98      Actinium
#> 91   0.0000000 0.00  0.73 1.00       Thorium
#> 92   0.0000000 0.00  0.63 1.00  Protactinium
#> 93   0.0000000 0.00  0.56 1.00       Uranium
#> 94   0.0000000 0.00  0.50 1.00     Neptunium
#> 95   0.0000000 0.00  0.42 1.00     Plutonium
#> 96   0.0000000 0.33  0.36 0.95     Americium
#> 97   0.0000000 0.47  0.36 0.89        Curium
#> 98   0.0000000 0.54  0.31 0.89     Berkelium
#> 99   0.0000000 0.63  0.21 0.83   Californium
#> 100  0.0000000 0.70  0.12 0.83   Einsteinium
#> 101  0.0000000 0.70  0.12 0.73       Fermium
#> 102  0.0000000 0.70  0.05 0.65   Mendelevium
#> 103  0.0000000 0.74  0.05 0.53      Nobelium
#> 104  0.0000000 0.78  0.00 0.40    Lawrencium
#> 105  0.0000000 0.80  0.00 0.35 Rutherfordium
#> 106  0.0000000 0.82  0.00 0.31       Dubnium
#> 107  0.0000000 0.85  0.00 0.27    Seaborgium
#> 108  0.0000000 0.88  0.00 0.22       Bohrium
#> 109  0.0000000 0.90  0.00 0.18       Hassium
#> 110  0.0000000 0.92  0.00 0.15    Meitnerium
#> 111  0.0000000 0.93  0.00 0.14  Darmstadtium
#> 112  0.0000000 0.94  0.00 0.13   Roentgenium
#> 113  0.0000000 0.95  0.00 0.12   Copernicium
#> 114  0.0000000 0.96  0.00 0.11     Ununtrium
#> 115  0.0000000 0.97  0.00 0.10   Ununquadium
#> 116  0.0000000 0.98  0.00 0.09   Ununpentium
#> 117  0.0000000 0.99  0.00 0.08    Ununhexium

# Get the mass of some elements
symb <- c("C","O","H")
elements[match(symb,elements[,"symb"]),"mass"]

#> [1] 12.01070 15.99940  1.00794

# Get the van der Waals radii of some elements
symb <- c("C","O","H")
elements[match(symb,elements[,"symb"]),"rvdw"]

#> [1] 1.70 1.52 1.10