Summarize and Print Features of a cna Network Graph

These functions attempt to summarize and print a cna network graph to the terminal in a human readable form.

# S3 method for cna
print(x, ...)
  # S3 method for cna
summary(object, verbose=TRUE, ...)

Arguments

x
A cna network and community object as obtained from the function ‘cna’.

object
A cna network and community object as obtained from the function ‘cna’.

verbose
Logical, if TRUE a community summary table is prited to screen.

...
Extra arguments passed to the ‘write.table’ function.

Details

Simple summary and print methods for protein dynamic networks.

Value

The function summary.cna returns a list with the following components:

id

A community number/identifier vector with an element for each community.

size

A numeric community size vector, with elements giving the number of nodes within each community.

members

A lst detailing the nodes within each community.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Guido Scarabelli and Barry Grant

See also

Examples

if (!requireNamespace("igraph", quietly = TRUE)) {
   message('Need igraph installed to run this example')
} else {

## Load the correlation network
attach(hivp)

## Read the starting PDB file to determine atom correspondence
pdbfile <- system.file("examples/hivp.pdb", package="bio3d")
pdb <- read.pdb(pdbfile)

## Examine network composition
print(net)
x<- summary(net)
x$members[[2]]

detach(hivp)

}

#> 
#> Call:
#>   cna.dccm(cij = cij)
#> 
#> Structure: 
#>  - NETWORK NODES#:   198 	EDGES#: 1756 
#>  - COMMUNITY NODES#: 18 	EDGES#: 13 
#> 
#>  + attr: network, communities, community.network,
#>         community.cij, cij, call
#>  id size                      members
#>   1   24      c(1:4, 94:107, 193:198)
#>   2    4                          5:8
#>   3   22               c(9:23, 64:70)
#>   4    4                        24:27
#>   5    2                        28:29
#>   6   49              c(30:63, 71:85)
#>   7    1                           86
#>   8    1                           87
#>   9    2                        88:89
#>  10    4                        90:93
#>  11   20 c(108:113, 117:122, 163:170)
#>  12   50 c(114:116, 129:162, 171:183)
#>  13    4                      123:126
#>  14    2                      127:128
#>  15    1                          184
#>  16    2                      185:186
#>  17    2                      187:188
#>  18    4                      189:192