Read mmCIF File

Read a Protein Data Bank (mmCIF) coordinate file.

read.cif(file, maxlines = -1, multi = FALSE,
         rm.insert = FALSE, rm.alt = TRUE, verbose = TRUE)

Arguments

file
a single element character vector containing the name of the mmCIF file to be read, or the four letter PDB identifier for online file access.

maxlines
the maximum number of lines to read before giving up with large files. By default if will read up to the end of input on the connection.

multi
logical, if TRUE multiple ATOM records are read for all models in multi-model files and their coordinates returned.

rm.insert
logical, if TRUE PDB insert records are ignored.

rm.alt
logical, if TRUE PDB alternate records are ignored.

verbose
print details of the reading process.

Details

The current version of read.cif reads only ATOM/HETATM records and creates a pdb object of the data.

See read.pdb for more info.

Value

Returns a list of class "pdb" with the following components:

atom

a data.frame containing all atomic coordinate ATOM and HETATM data, with a row per ATOM/HETATM and a column per record type. See below for details of the record type naming convention (useful for accessing columns).

xyz

a numeric matrix of class "xyz" containing the ATOM and HETATM coordinate data.

calpha

logical vector with length equal to nrow(atom) with TRUE values indicating a C-alpha “elety”.

call

the matched call.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Author

Lars Skjaerven

See also

Examples

## Read a mmCIF file from the RCSB online database
# cif <- read.cif("1hel")