Add a Sequence to an Existing Alignmnet

Add one or more sequences to an existing multiple alignment that you wish to keep intact.

seq2aln(seq2add, aln, id = "seq", file = "aln.fa", ...)

Arguments

seq2add
an sequence character vector or an alignment list object with id and ali components, similar to that generated by read.fasta and seqaln.

aln
an alignment list object with id and ali components, similar to that generated by read.fasta and seqaln.

id
a vector of sequence names to serve as sequence identifers.

file
name of ‘FASTA’ output file to which alignment should be written.

...
additional arguments passed to seqaln.

Details

This function calls the ‘MUSCLE’ program, to perform a profile profile alignment, which MUST BE INSTALLED on your system and in the search path for executables.

Value

A list with two components:

ali

an alignment character matrix with a row per sequence and a column per equivalent aminoacid/nucleotide.

id

sequence names as identifers.

References

Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

‘MUSCLE’ is the work of Edgar: Edgar (2004) Nuc. Acid. Res. 32, 1792--1797.

Full details of the ‘MUSCLE’ algorithm, along with download and installation instructions can be obtained from:
http://www.drive5.com/muscle.

Author

Barry Grant

Note

A system call is made to the ‘MUSCLE’ program, which must be installed on your system and in the search path for executables.

See also

Examples

if (FALSE) {
aa.1 <- pdbseq( read.pdb("1bg2") )
aa.2 <- pdbseq( read.pdb("3dc4") )
aa.3 <- pdbseq( read.pdb("1mkj") )

aln <- seqaln( seqbind(aa.1,aa.2) )

seq2aln(aa.3, aln)
}